java generation working
the way that the xslt runs is different again - it really does run against the model specified,
with the binding files specified as parameter.

Also, the filenames in the binding files should not contain any absolute path information, as
the XML catalogue mechanism is working to find the absolute location.