delete unnecessary file for gradle
Have deleted most of the auxiliary files that were required for the ant build, as all the ant functionality is now included
in the gradle tooling. The main ant build.xml file has been left as this might contain some important functionality that still
needs to be ported, however without the auxiliary files it is unlikely that it would run successfully.