initial minimal support for some composite primary key support
This is customized towards the needs of the TAPSchemaDM at the moment and should be generalised in future.

This is done by setting the natural key <position> indicator to 2 - heuristics in the code generation assume that the parent container key is needed as the first part of the composite key in the absence of a <position>1 natural key.